Tidy up test API handlers
